Creating server as a middleware
