+++ Back_End/app.js
@@ -1,35 +1,7 @@
-const express = require('express');
-const cors = require('cors');
-const app = express();
-const consign = require('consign');
 const dotenv = require('dotenv');
-
-
-// Definition of view engine and views directory
-app.set('view engine', 'ejs');
-app.set('views', './api/views');
-
-//Loads the environment variables
 dotenv.config()
 const port = process.env.PORT || 3050;
-
-app.use(cors());
-
-//Loads the middlewares to parse these data format
-app.use(express.urlencoded({
-    extended: true
-}));
-app.use(express.json())
-
-
-consign()
-    .include("./api/routes/authentication.js")
-    .then("./config/middleware.js")
-    .then("./api/routes")
-    .then("./api/models")
-    .then("./config")
-    .into(app)
-
+const app = require('./config/server')
 
 app.listen(port, function () {
     console.log("Server is running on port: " + port);
